requirements.txt
streamlit
pandas
matplotlib
numpy
plotly
pillow
fpdf
reportlab
kaleido